185-setup-warnings: do not install gnatcov_rts in gprbuild's prefix
Nothing in the whole testsuite is supposed to modify gprbuild's
installation prefix: in general the execution of tests must be
self-contained, without side effect for the environment.

This issue was found running the testsuite with the gnatcov_rts used by
the whole testsuite installed in gnatcov_rts: during the execution of
the 185-setup-warnings testcase, the installed project was temporarily
unavailable, causing the failure of the other tests running in parallel.

(cherry picked from commit 9a6fabfdd6c24f3cd73ab3808170d386ea13a5af)