gnatcov.gpr: remove the dependency on gnatcov_rts.gpr
This dependency is used only to pull a few data types and the version
number, but implies that the gnatcov codebase cannot be instrumented by
a gnatcov build based on an different coverage runtime. As a
consequence, all attempts to compute code coverage for gnatcov need to
compile gnatcov twice: once to get the "bootstrap" gnatcov, and once for
the instrumented gnatcov.

Without this dependency, one can use any gnatcov build to instrument the
gnatcov codebase. Also update the Makefile to add the
"--implicit-with=gnatcov_rts" gprbuild switch as expected for
instrumented builds.