instr.py: change default dump trigger for light-taskig runtimes
There seems to be an issue with the dependent fallback handlers of
Ada.Task_Termination in the light-tasking runtimes, resulting in the
handler not being executed at the end of the main. This thus makes
it impossible to use that dump trigger, so change the default dump
trigger of light-tasking runtime to be main-end.