Introduce the "gcc-wrapper" internal subcommand
Move the actual integrated instrumentation implementation to this new
subcommand so that the compiler wrapper is just a thin wrapper around
gnatcov itself. Thanks to this, the compiler_wrapper-gcc main does not
need to pull the C instrumenter anymore: it is faster to compile/link
and yields a smaller executable.