Setup: Refine the heuristic for full runtime detection
The QNX light-tasking runtime provides the Ada.Command_Line unit, but is
not a full runtime (as it does not provide e.g. GNAT.OS_Lib). Check for
the presence of GNAT.OS_Lib rather than Ada.Command_Line to determine
whether the runtime is a full runtime or not.