.gitlab-ci.yaml: Use the standard e3 image for all jobs
The gnatcov image was used to get a pre-built binutils in the
sandbox. This is no longer required now that binutils is a component
and can thus be downloaded.

This also brings all the generic_anod_ci capabilities to the gnatcov
repo, including arbitrary depends-on, release branch support and so
on.