instr-cov/198-non-instr-scos: kill with 5.04a1 toolchains
This testcase relies on an Ada 2012 construct (expression function) in
order to trigger a "uninstrumented" case in gnatcov: there is no
pre-2012 construct that is known to have this effect. So just disable
this testcase with toolchains that do not support Ada 2012.

(cherry picked from commit 221f260001d1ea5bd55c083876e121c19a256742)