Treat coverage runtime loading failure as a fatal error
Attempts to load the coverage runtime in gnatcov were so far optional:
in case of failure (gnatcov_rts.gpr cannot be found, wrong
target/RTS is used, ...) some output was produced in verbose mode and
processing continued with default settings instead of using the
configuration that was supposed to be loaded from the coverage runtime.

This proved to be a burden for users, as issues that could be detected
at "gnatcov instrument" time were invisible, and caused hard to diagnose
problems later on (for instance when building instrumented programs). To
fix this, make gnatcov abort when it cannot successfully extract
information from the coverage runtime project.