Bump trace format version
The trace entries used to contain the name of the project to which
the trace entry's unit belonged. This was removed by
b567aca989fa1a718f80987d9c45e5dfb4bb1839, but the change did not
bump the trace format version despite being a breaking change.

This change bumps the version number after the fact to avoid potential
confusing error messages in the future.