GNATcov_RTS: remove uses of Ada.Command_Line.Command_Name
Passing the actual command name brings little value: the program name
recorded in the source trace is not used in practice, and passing the
name of the project file instead looks good enough.

This change goes in the way of avoiding dependencies on the secondary
stack in instrumented projects, which may have the No_Secondary_Stack
restriction.