Instrument Ada: fix the handling of configuration pragmas files
After the work on parallelized intrumentation, configuration pragmas
mappings were tentatively passed from the "instrumentation driver" to
the "instrumentation subprocess" using the file that the
GNATCOLL.Projects.Aux.Create_Config_Pragmas_File function creates. The
problem is that this function is not about the same "configuration
pragmas": it seems to process pragmas that deal with source file
mappings. As a consequence, the handling of configuration pragmas files
was completely broken.

Switch back to a pure Libadalang-based implementation of configuration
pragmas mappings to restore the previous behavior, and add a (non
XFAILed) testcase to check it works as expected.