Manual dump: Emit one helper unit per project per language
Gnatcov used to only emit a single helper unit per project, either
in Ada or in C, which effectivelly prevented the use of the feature
when dump indications were contained both in C-like and Ada sources.

Instead of tracking for which project a helper unit has already been
generated, track instead the set of helper units which have been
genereated. This way if two units in the same project require different
helper units (because they aren't in the same language), both will get
emitted.