GNATcov_RTS: do not use Interfaces.C.Strings.Value
This is a function that returns a string, and thus its usage necessarily
implies a dependency on the secondary stack. Since instrumented projects
may be subject to the No_Secondary_Stack restriction, this means that we
cannot use thus fonction in the coverage runtime.

(cherry picked from commit 3967790525aadf0948727f2440d015f74f2493bb)