instr-cov/mixed_manual_dump: XFAIL test for cross configurations
There's an inconsistency in the order of object files / library files
on the command line generated by gprbuild depending on whether there
is only a C main in the projet or both C and Ada mains.

In the later case (this test), our last chance handler setup fails to
properly build.