Add buffer reset mechanism for source traces
This is based on the same mechanism than the one for the manual dump
trigger. Users can add

   `pragma Annotate (Xcov, Reset_Buffers);`

or

  `/* GNATCOV_RESET_BUFFERS */`

depending on the language of the source to clear all the coverage
buffers for all the units in the project tree rooted at the project
in which the unit with the reset indication belongs.