instr-cov/208-no-secsta: rework to harmonize src/bin traces
At higher optimization levels, the compiler is able to eliminate dead
code in main.adb, and this affects the coverage report in binary traces.

Move some code to a separate unit to avoid inlining and thus disable the
problematic dead code elimination. This fixes discrepancies in coverage
reports with bin/src traces.