Project.Load_Root_Project: avoid null/empty string confusion
Turn the DB_Dir format into a string, so that there is only one value
that means "no db dir": the empty string. So far, Load_Root_Project
considered that only a null string access value meant that, whereas the
caller used an empty string instead.

As a consequence, GNATCOLL.Projects.Add_Config_Dir was always called
(generally with an empty string), which could cause the GPR system to
fail to detect the C compiler and trigger spurious "is not a file or
directory" error messages from gprconfig.