Integrated instrumentation: fix the computation of the buffers list
To compute the list of buffers in the link closure in the integrated
instrumentation mode, gnatcov searches all buffer-like symbols in the
object / library / source files referenced in the link command line.
This has the downside of pulling all instrumented units from a static
library, which can result in a link failure.

To fix this, we link two times:

  * The first time, we link against an empty buffers list.

  * The second time, we link against a buffers list unit that references
     exactly the list of coverage buffers included in the first link.