instrument-project.adb: regroup context cleanup to context destruction
Cleanup for temporary files used by the insturmenters was done
after finishing the coverage instrumentation, thus some artifacts
were missing during manual dump instrumentation.

This change moves the cleanup of temporary files to the destruction
of the instrumentation context, after which point it is sure that the
files will not be used anymore.