instrument-ada_unit.adb: Only query a LAL unit once for each file
The manual dump indication processing used to query a LAL unit
twice when processing a file with manual dump / reset indications:
once for the tree traversal, and once when starting the rewrinting
session when the manual indication was found. If the LAL context was
reset on the second query, the instrumentation would fail due to stale
references.

This change adds a mechanism to start a rewriting session from an already
existing analysis unit.