SC_Obligations: rework scope traversal
The previous scope traversal algorithm relied on forward iterators
for multiway trees, which was an issue when there were SCO gaps in
the traversal. It also imposed that the traversal had to be in
increasing SCO order.

The traversal is now done manually, which means it can be done in
arbitrary order (although traversing with SCOs in ascending ordered
will increase performance).