Instrument: fix consolidation support for stubbed subunits
In the case of stubbed subunits, the SCOs fingerprint can be identical
for each "unit" (CU_Info tense) between instrumentations of the original
sources and stubbed projects, but the bit maps will be slightly
different in a way that would not hurt consolidation; as a consequence,
before this commit, bit maps fingerprints will differ and so gnatcov
will refuse the consolidation.

An example to clarify:

  package body Pkg is
    procedure Proc_A is separate;
    procedure Proc_B is separate;
  end Pkg;

  --  Real implementations for Proc_A and Proc_B:

  separate (Pkg)
  procedure Proc_A is
  begin
     Stmt_1;
     Stmt_2;
  end Proc_A;

  separate (Pkg)
  procedure Proc_B is
  begin
     Stmt_3;
     Stmt_4;
  end Proc_B;

The instrumentation of original sources will yield the following
low-level SCOs, high-level SCOs and bit maps:

  LL SCOs:
    For pkg-proc_a.adb:
      SCO for Stmt_1
      SCO for Stmt_2
    For pkg-proc_b.adb:
      SCO for Stmt_3
      SCO for Stmt_4
  HL SCOs:
    For pkg-proc_a.adb:
      SCO #1 for Stmt_1
      SCO #2 for Stmt_2
    For pkg-proc_b.adb:
      SCO #3 for Stmt_3
      SCO #4 for Stmt_4
  Bit maps (for statements):
    For pkg-proc_a.adb:
      1 => SCO #1
      2 => SCO #2
    For pkg-proc_b.adb:
      1 => SCO #3
      2 => SCO #4

However for a build in which Proc_A is stubbed (and thus ignored with
--ignore-source-files), we will get instead:

  LL SCOs:
    For pkg-proc_b.adb:
      SCO for Stmt_3
      SCO for Stmt_4
  HL SCOs:
    For pkg-proc_b.adb:
      SCO #1 for Stmt_3
      SCO #2 for Stmt_4
  Bit maps (for statements):
    For pkg-proc_b.adb:
      1 => SCO #1
      2 => SCO #2

The SCOs fingerprint is computed from low-level SCOs: stubbing unit X
does not affect the fingerprint of SCOs for unit Y, so we are good.
However, currently bit maps fingerprint is computed using high-level
SCOs: stubbing unit X will change the numbering of these SCOs, and will
thus affect the bit map, and thus its fingerprint: gnatcov will refuse
to consolidate source traces created from stubbed builds when it has
loaded SIDs created on original sources because of this bit maps
fingerprint mismatch.

This commit fixs the way we compute fingerprints for instrumentation bit
maps so that the fingerprint is identical for Pkg regardless of
whether/how the Proc_A/Proc_B subunits are stubbed: instead of using
"absolute" SCO ids, we now use SCO ids that are relative to the
corresponding unit (i.e. a SCO "offset" from the first SCO of the
current unit). Offsets for unit Y do not vary depending on the stubbing
of X, and they still help reflecting the structure of bit maps: the
fingerprint still plays its safeguard role.

(cherry picked from commit 1a4a491b392f0f51c8bd09fd8f0d6d053f359a89)