Fix crash upon missing exemption justification
gnatcov used to expect the justification to always be filled in the
pragma signaling the start of an exempted region. This led to null
pointer access when it was missing.

Now, a warning is emitted during instrumentation and the justification
is replaced with "No justification" in the report.

Similarly, any deviation from the exact expected pragmas is now be reported
as a warning.