rts/gnatcov_rts-traces-output-files.adb: kill warnings for no exc. prop.
Extend the existing workaround in gnatcov_rts to accomodate for the
runtimes that have the No_Exception_Propagation restriction to this
source file: native runtimes are unlikely to have this restriction, but
gnatcov_rts's full profile is actually compiled with some non-native
runtimes (for instance: aarch64-qnx/light-tasking).

A commit following this one will add a testcase to check general support
for the full profile on such runtimes.