SUITE.tutils.run_cov_program: handle dump-channel=bin-file with pycross
On targets that have filesystem support, it is possible for testcases to
use the bin-file dump channel, and thus have the source trace file
created directly on the target filesystem. pycross happens to mount that
filesystem on the host, and with run-cross2's --save-temps argument, it
is possible to access that mount point after the target program has
completed. Do this to extract source traces and put them where the rest
of the testtsuite expect them: in the current directory.

(cherry picked from commit 7c209dcf551b23b1085ce88df946e1ec7ccf5162)