instr-cov/W502-023-base64-long-lines: stabilize for arm-elf
Something in the pipeline (the Ada runtime? Qemu?) has trouble handling
the very long lines that the test program writes on the serial port,
making this test unstable on arm-elf.

Since the point of this testcase is to make sure that "gnatcov
extract-base64-trace" does not crash (regardless of how the input text
file was created), rework it so that the very long line is written by
Python, which should fix the unstability.