Use dominant information to refine coverage data
We use the dominant information from statements that are covered by a
program execution, but we do not use this information when coverable
statements with dominant information are not covered. This information
can actually be used to set a known outcome taken for a given decision,
and to not have degraded origins which can result in uncomplete coverage
results being reported at e.g. consolidation.