Force references to coverage buffers in Ada mains for manual dump trigger
When using a manual dump trigger, with sources of interest in the root
project, but no dump indication in that project, the coverage buffers
units for that project would then not be in the main's compilation closure,
resulting in a link error.

This change fixes this by always making a reference to the dump helper unit
in the Ada mains, which transitively make references to all the buffer units
for the whole project tree.