Prevent local subprogram optimization on Qualif/.../Pragmas/Policy_Off
A subprogram only called through "pragma Debug" statements
is never actually called when testing with the pragma Assertion_Policy
off.

When the subprogram is local to a unit, optimization just gets rid
of the corresponding code entirely and the corresponding statements
are reported "no code" with binary traces.

This change just exposes the said subprogram to prevent the removal,
allowing consistent expectations across trace and optimization modes.