Attach violations to last location
All violations were always attached to the first line of the statement
or decision it is relative to. For report formats displaying violations
in the source code, whenever a statement or decision spaned multiple
lines it was "cut" by the violations as they were shown directly under
the first line.

To improve readability, violations are now attached to the last line of
the statement or decision so they can be displayed after the last line.
This also has the added bonus of allowing the user to select the text of
the statement or decision with no hindrance from the violations.