Instrument.Projects: Remove spurious directory check
When using the manual dump trigger, gnatcov used to check,
for all projects, that there exists a directory with the same basename
as the project's object directory. This is not needed as we also check
the existence of the output dir (which is inside the object directory),
and breaks the feature when gnatcov is not invoked exactly in the same
directory as where the object dir will be created.

This change removes the spurrious check.