Only use the last occurence of --dump-trigger
--dump-trigger used to be a string option, which meant only the last
occurrence was used, and previous ones were ignored. This changed when
--dump-trigger=manual started accepting a comma-separated list of files
in which to search for a manual dump indication, and thus all occurrences
of the --dump-trigger switch were concatenated together.

This change restores the original behavior, only taking into account the
last --dump-trigger switch, but still processing the value as a
comma-separated list.