Consider decisions with no object level control flow impact as non-coverable
These are decisions with no branches in the executable and which do
not dominate any statement (such as if-expressions).

As their evaluation has no impact on the execution of the program,
and their valuation cannot be traced back from other statements, consider
that there is nothing to be covered.

Also modify the testsuite to recognise non-coverable decisions.