Use the most exending project for manual dump trigger processing
When processing files for a manual dump trigger, gnatcov used to iterate over all
the sources in the project tree. Each source file is associated to the project
in which it is originally located by GNATCOLL.Projects, whereas we would need
to associate each file to the most extending project for its original project,
to ensure the coverage buffers for all files in the extending project are dumped,
not only those in the extended project.

This change ensures that each file is associated to the most-extending project
when processing it for manual dump trigger items.