Add new coverage disabling annotations
This commit introduces Cov_Off / Cov_On annotation to selectively
disable coverage analysis over a source code region. This is done
through pragma Annotate for Ada, and comment markers for C.

The disabled coverage regions are reported as D lines (for Disabled) in
the various kind of reports, and have there is a dedicated section for
the report format.