Modify driver behavior when using --consolidate=checkpoints
When using the --consolidate=checkpoints testsuite option, the driver
will systematically produce a checkpoint, and then produce coverage
reports (xcov, and report) from that checkpoint, even when producing a
coverage report for a single test driver.

This commit is in the context of adding support for disabling coverage
annotations, which will only be displayed as such when producing a
report by using checkpoints.