Merge annotations from different checkpoints
This commit is in the context of introducing support for coverage
disabling annotations. It is allowed to have a varying set of coverage
disabling annotations for two different checkpoints, and we want to
account for all of them in the resulting report.

As the previous behavior was to ignore the annotations from a checkpoint
if the compilation unit had already been loaded from another checkpoint,
we change it to now also include the annotations from the newly loaded
checkpoint.