Instrument.C: fix handling of manual dump/reset indications
The current logic to handle both kinds of indication gets stuck in an
infinite loop when the two last indications are a dump, then a reset.
Rework this area to use a single regular expression to cover both kinds
of indication, which simplifies the handling loop and avoids the faulty
logic.