Add new commands to manipulate external annotation files
The three new commands, add-annotation, delete-annotation and
show-annotations allow the user to generate and inspect stable_sloc
entry files, without needing to know the syntax nor the expected
fields for the annotations recognized by gnatcov.

A basic test is added in this commit, more thorough testing will be
added in the following commit to simplify the review process.