Add new tests for external annotation generation
The new tests are copies of pre-existing tests concerning in-source
annotations, that have been modified to neutralize the existing
annotations. Instead, invocation of "gnatcov add-annotation" are used
prior to the original test script to generate external equivalent
annotations. The results are supposed to be identical.

This has been split as a separate commit from the previous one to
facilitate review.