Instrument.C: Fix handling of escape sequences in preprocessor output
gnatcov interprets line directives emitted by the preprocessor to relocate
external annotations. This processing did not account for escape sequences
generated by the preprocessor, in particular for escaped '\' characters
when the host is Windows, resulting in all buffer annotations being ignored.

This also fixes an infinite loop which only appears when some annotations could
not be remapped.