Coverage.Source: Fix processing of block coverage
The introduction of function and call coverage added new SCO types
that are discharged by bits in the statement buffer, but which were
not part of any block.

This change properly ignores the buffer bits that are not part of any
block when processing the block information at coverage time.