instrument-ada_unit.adb: fix wrong use of Insertion_Info_Ref
Some parts of the instrumentation used to make a copy of the referenced
insertion info in the instrumentation context when processing nested entites,
thus losing the reference to the insertion information referenced by the
block stack. This resulted in access to invalid insertion information.

Now only references to insertion info are saved / restored when processing
nested entities, thus ensuring proper preservation of the insertion
information across the instrumentation context, and the block stack.