Fix handling of exemption violations for report + another format
Exemption violations are handled differently when generating a "report"
format. To achieve this, conditional code checks this case with a
"not Annotation (Annotate_Report)" condition.

That condition was fine at the time when "gnatcov coverage" could
generate only one annotation format at a time, however nowadays a single
invocation can generate multiple annotation formats, so multiple
booleans in the "Annotation" array can be set to True simultaneously:
"Annotation (Annotate_Report)" may be True even when we are generating a
non-"report" format such as "xcov". As a consequence, code that is meant
to be executed only when generating a non-"report" format is skipped at
the wrong time, and so exemption violations are incorrectly processed.

This commit introduces a new Pretty_Printer primitive so that
conditional code can check the format that is currently generated
(instead of the formats that will eventually be generated throughout the
execution of "gnatcov coverage") and uses it to fix the handling of
exemption violations when multiple reports are generated.

(cherry picked from commit f33a030cec108ae09bd027c932956355c44c0657)