Simplify function coverage for subprogram bodies
Before, function coverage was performed for subprogram bodies the same
way as for expression functions: by adding an auxiliary function
responsible for calling the original one. This is more complex than it
could be.
Now, function coverage for subprogram bodies is done the same way as for
generic procedures: as the presence of a declarative part is guaranteed,
the witness call respondible for discharging the function SCO is
inserted in a dummy variable declaration at the very beginning of the
subprogram's declarations.