GNATcov_RTS: Add new configuration suitable for a -nostdlib config
This adds a new "minimal" profile to GNATcov_RTS, which makes it possible
to link an instrumented project with no Ada runtime library.

The change includes a new implementation for gnatcov_rts_put_string, relying
on an externally defined putchar-like function to be provided by the user, and
some configuration logic to properly select the right unit based on the external
variables defined.

gnatcov setup has also been enhanced to expose the minimal profile, and configuration
options to the user.