M528-002-transient-finalize/alternate: cope with return copy elision
Recent GNAT improvements allow the compiler to elide copy in aggregate
returns for nonlimited types. As a result:

   return (Controlled with V);

No longer triggers a call to the Adjust primitive for Witness. Break
this into:

   Result : constant Witness := (Contolled with V);
   return Result;

To make sure that a copy is made, and thus to make sure that the Adjust
primitive is called.