rundrv-handlers.adb: restore -a switch for visium-elf-run cmd
This parameter was removed in 9ab22d33 due to it no longer being
required in gdbsim. The only gdbsim build we maintain for visium-elf
is on the 7.2.x setup, which predates that change, so we need to
restore the -a parameter in order for gnatcov run to work for that
setup.