Pass the buffers list object instead of source file
Instead of passing the source file version of the buffers list unit,
compile it and then pass the resulting object file. In the former case,
we end up reusing compiler switches from the compiler driver original
command. This is a twofold issue, as these switches are not needed to
compile the buffers list unit, and it can result in symbols clashing
when e.g. the user passed a `-include` switch on the compiler driver
command line.