instr-cov/339-nostdlib: Use light runtime in all gnatcov invocations
This ensures that gnatcov setup does not attempt to build a shared
library which results in a link failure on windows.

Building an instrumented application with a non-light runtime but with the
-nostdlib switch doesn't really make any sense, as this would result in
references to the runtime (e.g. through elaboration of the non-light versions
of the runtime) so not testing this configuration isn't an issue in practice.