SCOV: Add option in CovControl to override dump trigger choice
This adds a new member to CovControl, allowing to override the default
dump trigger choice for tests run in source trace mode.

The driver is adapted to look for the correct trace pattern for all
dump triggers.

The last chance handler hook insertion is also adapted to account
for the differences in helper unit naming pattern when a manual
dump trigger is used.