instr-cov/326-pragma-loc-check: various adaptations
The trace prefix set in the main source interferes with the testsuite
mechanism to find the correct trace after execution. Remove it as it
is not required for the test to work properly.

The test script now uses the new CovControl option to set the dump
trigger, instead of relying on the instroption to force the dump trigger.