Fix XML report production for non-UTF-8 source code
Consider by default that sources are encoded in Latin-1. For the cases
they are not, also add a --source-encoding option to allow users to
specify which encoding to use.

The adaptation of SCO_Text to handle encodings fixed bugs in the
truncation logic as a side effect. Adjust the default truncation length
to reflect what was implemented (length = 9 instead of 8) to limit the
need to adjust baselines.

(cherry picked from commit d247432a47a885b52cad429f2b8594ab718f7021)