sc_obligations.adb: fix assertion for the case of SID/checkpoint loading
When a SID or checkpoint file is loaded, the coverage obligations it
holds are merged into global tables in the Checkpoint_Load_Merge_Unit
procedure. This process eventually calls the Condition function (to
compute the SCO_Id for a condition from its enclosing decision + its
index), which itself calls the Index function (to get the index of a
given SCO condition) in an assertion.

While the Condition function accepts a Source_Coverage_Vectors argument
to operate either on global tables or on tables loaded from a
SID/checpoint file, the Index function exclusively work on global
tables. As a consequence, in the context of SID/checkpoint loading,
the call to Index is meant to operate on SCOs from "loaded tables" but
in fact operate on global tables, i.e. on SCOs that may be completely
irrelevant.

This commit introduces the necessary Index overload for it to be able to
work on custom Source_Coverage_Vectors tables and fixes Condition to use
that overload. This fixes a crash in
TA27-015-ignore-source-files-source-coverage, found during the
transition to GPR2.

This looks really bad and one might wonder why this was not found
before: this code path is exclusive to the processing of decisions whose
valuation is statically known, and the global/loaded table disconnect
that is needed to trigger the crash, which seems to be extremely rare in
practice. The GPR2 transition revealed this issue probably because it
changed the order of SID processing, which affects when the
global/loaded table disconnect happens.