Instrument.C: Do not output clang diagnostics as warnings
We currently do not have a good heuristic to determine when a parsing
diagnostic emitted by libclang is actually a concern, so diagnostics are
suppressed by default. Until now, they were emitted as warnings, only if
verbose logging was activated, and contained the severity of the diagnostic
text (e.g. error or fatal error), which could be confusing for the user, as
it lead to believe there is an error when it is most often not the case.

The diagnostic emmission mechanism has been reworked to clarify that the
messages are not generated by gnatcov directly, and the severity has been
removed.

No behavior change expected.