Instrument.Ada_Unit: Propagate calling convention to null procedure wrappers
Null procedures are instrumented by instantiating a generic procedure to host
the witness call. This instance should be subtype conformant with the procedure
that renames it, according to RM 8.5.4 (5/3). This includes matching calling
conventions, the calling convention of the original null procedure must thus also
be applied to the instance.