Instrument.C: Fix pre-processing of sources where comments cannot be preserved
Some macros can perform token concatenation, but do not allow
concatenating comments with other tokens. The comments are normally replaced
by white         spaces before macro expansion, but gnatcov needs to preserve comments
to process annotations. In this case the preprocessor will error out when asked
to preserve the comments.

To fix this, this change introduces a retry without the -C argument. This allows
gnatcov to properly instrument the source, but annotations will be ignored.