C Instrumentation: Detect when the compiler is native 32bit
The preprocessing of sources and libclang need to use the same set
of option when processing sources, which is not the case when using
a native x86 compiler, as libclang is not aware that it needs to process
the sources as if targeting a 32bit executable.

This change adds an explicit -m32 switch to the compiler command lines when
detecting that the compiler targets i686-pc-linux-gnu. We may need to make this
heuristic a bit broader in the future if we also have issues with embedded 32bit
targets (e.g. riscv32-elf or arm-elf).