Instrument.C: Fix filename computation for 32bit check
The filename used to include the full compiler driver name, which
could be a full path, resulting in writes in a directory where we
potentially did not have permissions.

Only used the compiler simple name instead, and ensure the output
file is created in a directory where we can put artifacts.