Doc: Clarify Ada runtime dependency for the minimal runtime profile
The minimal runtime profile does not depend on any function or object
from the Ada runtime, but runtime checks generate a reference on the
last chance handler symbol. On light runtime, the default implementation
uses the IO units to output an error message, thus forcing dependencies
on some runtime units.