gnatcov_llvm_json: statically link libstdc++ and libgcc
This aligns the build of `gnatcov_llvm_json` with the builds of the rest
of our tools. Doing so is necessary for this trace adapter to work in
contexts where the environment does not provide these shared libraries,
or when they provide outdated ones (for instance when using GNAT 7.1.2,
which provides a fairly old libstdc++).