TC03-012-objcov-line-state: retarget for AArch64 ELF
This testcase for object coverage (binary traces) is hardwired for
x86_64-linux. Since we no longer support binary traces for that
platform, this test never runs in practice. Retarget it for AArch64 ELF
so that we can run it again.

The original problem (instruction reported as covered even when not
executed) reproduces if the fix in `coverage`-object.adb is reverted.