instr-cov/C++/std-cond-code: fix with optimizations enabled
When optimizations are enabled, the call to `print_msg` is inlined, and
thus the string literal (unused in `print_msg`) is optimized aways, not
present in the output assembly code. This testcase relies on that string
literal to perform its check.

Move `print_msg` to a separate source file to avoid inlining, and thus
this optimization.