Instrument Ada: fix instrumentation of decisions in formal decls
Because default expressions in formal declarations are executed only
when their owning subprogram is called (instead of during the
elaboration of that subprogram), it is incorrect to use the declarative
region that contains the subprogram in order to insert MC/DC state
variables for decisions in default expressions. Report a gnatcov
limitation instead in that case.

In addition, subprogram spec can be duplicated (one occurrence for a
subprogram declaration, another for the body). Right not it is not
possible to use the same instrumentation (same bit in buffer coverage)
for both, so decision coverage is also impossible.

The existing (disabled) test for this is repurposed to check this case.