Instrument Ada: fix instrumentation of decisions in component decls
Because default expressions in component declarations are executed only
when a record object/aggregate is declared (instead of during the
elaboration of the record type), it is incorrect to use the declarative
region that contains the record type declaration in order to insert
MC/DC state variables for decisions in default expressions. Report a
gnatcov limitation instead in that case.

The existing (disabled) test for this is repurposed to check this case.