Subprocesses: ensure no process handle is left uninitialized
Use the explicit constant Invalid_Handle for free slots in the `Handles`
process handle array (both at pool initialization and when a process
terminates).

This will avoid unexpectedly waiting for a process that is not part of
the pool (rare, but possible, if a PID gets reused on the host, or if
uninitialized slot contains a valid PID by chance).