Deduplicate the gnatcov_rts.ads source
It contains among other things the current runtime version, and was
until now duplicated between the regular runtime and the Ada only
version. To avoid forgetting to update one file and not the other
in the future, only keep the one from the regular runtime and copy it
when installing gnatcov.