Revert "Merge branch 'eyraud/429' into 'master'"
This reverts commit 69a14415abc314a779c4c8653dc1f578b7b4eed3, reversing
changes made to 3341e2fe68474dfbff82916ecf9902dcb35b06b1.

The issue is that this MR localized all coverage obligations from an included
header on the single line where the inclusion takes place, and on the column
range corresponding to the buffer index in the original header.

When creating low leve SCOs, the column number is limited to 2^15, which is easily
smaller than the number of characters in a header.

Ignoring these SCOs reveals more challenges, so this is reverted while we take some
time to find a better encoding for these SCOs.