Port raise expression tests to the Qualif testsuite
The current tests have been added to the mcdc/1_core/NoEval as they
help demonstrate how gnatcoverage correctly reports the decision as
not evaluated as soon as the raise expression is evaluated.

A new simpler test has been added to decision/1_core/NoEval to demonstrate
that gnatcov correctly reports the decision as unevaluated when no outcome
is ever reached due to the raise expression, but that its evaluation in one
decision eval does not affect coverage of other evaluations.